Module gawireslib::cli::subcommand

Structs

Add

track new assets or changes, add tags and other operations.

Branch

Operations on branches

Central

Commands related to the creation or management of a Gawires Central Server.

Channel

Create and manage channels, channels are used to have specific assets versions or representation of a same project.

Checkout

Checkout assets in local workspace. When assets are checkout, they are locked by default when in centralized mode.

Clean

Cleanup tools for local workspace

Connect

Connect and sync a Centralized Workspace.

Filter

Filter operations

Gist

Operations over local or remote asset’s gist

Init

Initialize a new project workspace

Install

Install a Gawire extensions.

Merge

Merge changes between 2 branches or centralized workspaces

Preview

Preview and compare different types of asset

Pull

Used in a distributed workspace to pull localy changes from remote.

Push

Used in a distributed workspace to push local changes to remote.

Rebase

Rebase a branch or a centralized workspace

Release

Release checkout assets from local workspace.

Remove

Unstage new assets or changes

Reset

Reset workspace assets to an earlier state

Status

Log and fetch information about local workspace

Sync

While used in a distributed workspace this will fetch the remote, while in centralized workspace this will synchronize local workspace state with remote.

Tag

Add or remove tags from assets or project version

Uninstall

Uninstall a Gawire extensions.

Wire

Create or configure wires.

Workspace

Commands to manage workspace

Enums

Subcommand

Enum of all possible subcommands